Update iOS deployment target in podspec to 8.0
The podspec claimed support for iOS 4.0, but the source code makes use of methods only available in iOS 6 or later (for example, -objectForKeyedSubscript:). The deployment in the Xcode project was 8.0, so the podspec deployment target is now also 8.0. This will avoid a lot of warning when the pod is used as a dependency.